+++ examples/products-list.php
@@ -3,10 +3,11 @@
 declare(strict_types=1);
 
 use Aceproxies\ResellerApi\Client;
+use Aceproxies\ResellerApi\Enum\ProductType;
 
 /** @var Client $client */
 $client = require __DIR__ . '/_bootstrap.php';
 
-$products = $client->products()->list('residential_proxy');
+$products = $client->products()->list(ProductType::ResidentialProxy);
 
 var_dump($products);